Refactor RomaApplication to be more readable
